Moved ShellSubHeading component to UI, to prevent recursive imports from shell component
