+++ packages/ui/components/layout/index.ts
@@ -1,0 +1,1 @@
+export { ShellSubHeading } from "./ShellSubHeading";
